sorting_iter: use FusedIterator instead of manual implementation
Reduces generated assembly by about 85%.

Before this, when an `Iterator` exhausted we would replace it with
a `None`, in order to be sure not to call `Iterator::next` ever again.
This is because some iterators don't like being called again after
they have been exhausted.

This manual implementation wasn't necessary, as `FusedIterator` allows
to achieve this exact task. Also, fusing an iterator that implements
`FusedIterator` is a no-op, since the underlying `Iterator` already correctly
handles being called again after exhausting it's items.

More info on `FusedIterator`:
https://doc.rust-lang.org/1.48.0/std/iter/trait.FusedIterator.html

Signed-off-by: Paolo Barbolini <paolo@paolo565.org>